+++ native/c/commands/tool.cpp
@@ -98,45 +98,6 @@ int handle_tool_dynalloc(InvocationContext &context)
   return rc;
 }
 
-int handle_tool_display_symbol(InvocationContext &context)
-{
-  int rc = 0;
-  std::string symbol = context.get<std::string>("symbol", "");
-  std::transform(symbol.begin(), symbol.end(), symbol.begin(), ::toupper);
-  symbol = "&" + symbol;
-  std::string value;
-  rc = zut_substitute_symbol(symbol, value);
-  if (0 != rc)
-  {
-    context.error_stream() << "Error: asasymbf with parm '" << symbol << "' rc: '" << rc << "'" << std::endl;
-    return RTNCD_FAILURE;
-  }
-  context.output_stream() << value << std::endl;
-
-  return RTNCD_SUCCESS;
-}
-
-int handle_tool_list_parmlib(InvocationContext &context)
-{
-  int rc = 0;
-  ZDIAG diag{};
-  std::vector<std::string> parmlibs;
-  rc = zut_list_parmlib(diag, parmlibs);
-  if (0 != rc)
-  {
-    context.error_stream() << "Error: could not list parmlibs rc: '" << rc << "'" << std::endl;
-    context.error_stream() << "  Details: " << diag.e_msg << std::endl;
-    return RTNCD_FAILURE;
-  }
-
-  for (const auto &lib : parmlibs)
-  {
-    context.output_stream() << lib << std::endl;
-  }
-
-  return rc;
-}
-
 int handle_tool_search(InvocationContext &context)
 {
   int rc = 0;
@@ -451,17 +412,6 @@ void register_commands(parser::Command &root_command)
   tool_dynalloc_cmd->add_positional_arg("parm", "dynalloc parm string", ArgType_Single, false);
   tool_dynalloc_cmd->set_handler(handle_tool_dynalloc);
   tool_cmd->add_command(tool_dynalloc_cmd);
-
-  // Display symbol subcommand
-  auto tool_display_symbol_cmd = command_ptr(new Command("display-symbol", "display system symbol"));
-  tool_display_symbol_cmd->add_positional_arg("symbol", "symbol to display", ArgType_Single, true);
-  tool_display_symbol_cmd->set_handler(handle_tool_display_symbol);
-  tool_cmd->add_command(tool_display_symbol_cmd);
-
-  // List-parmlib subcommand
-  auto tool_list_parmlib_cmd = command_ptr(new Command("list-parmlib", "list parmlib"));
-  tool_list_parmlib_cmd->set_handler(handle_tool_list_parmlib);
-  tool_cmd->add_command(tool_list_parmlib_cmd);
 
   // Search subcommand
   auto tool_search_cmd = command_ptr(new Command("search", "search members for string with parms, e.g. --parms anyc"));